+++ source/braille/__init__.py
@@ -42,11 +42,11 @@ def terminate():
 __getattr__ = handleDeprecations(
 	# MovedSymbol("BrailleDisplayDriver", "braille.display.driver"),
 	# MovedSymbol("BrailleDisplayGesture", "braille.display.gesture"),
-	MovedSymbol("getSerialPorts", "braille.display"),
-	MovedSymbol("getDisplayList", "braille.display"),
-	MovedSymbol("getDisplayDrivers", "braille.display"),
-	MovedSymbol("RENAMED_DRIVERS", "braille.display"),
-	MovedSymbol("DisplayDimensions", "braille.display"),
+	# MovedSymbol("getSerialPorts", "braille.display"),
+	# MovedSymbol("getDisplayList", "braille.display"),
+	# MovedSymbol("getDisplayDrivers", "braille.display"),
+	# MovedSymbol("RENAMED_DRIVERS", "braille.display"),
+	# MovedSymbol("DisplayDimensions", "braille.display"),
 	MovedSymbol("Region", "braille.regions.base"),
 	MovedSymbol("RegionWithPositions", "braille.regions.base"),
 	MovedSymbol("TextRegion", "braille.regions.base"),
