+++ source/brailleDisplayDrivers/baum.py
@@ -7,6 +7,7 @@ from io import BytesIO
 from typing import Union, List, Optional
 
 import braille
+import braille.display
 import braille.display.driver
 import braille.display.gesture
 from hwIo import intToByte, boolToByte
@@ -156,7 +157,7 @@ class BrailleDisplayDriver(braille.display.driver.BrailleDisplayDriver):
 
 	@classmethod
 	def getManualPorts(cls):
-		return braille.getSerialPorts()
+		return braille.display.getSerialPorts()
 
 	def __init__(self, port="auto"):
 		super(BrailleDisplayDriver, self).__init__()
