+++ source/brailleDisplayDrivers/brailleNote.py
@@ -14,6 +14,7 @@ from typing import List, Optional
 import serial
 import bdDetect
 import braille
+import braille.display
 import braille.display.driver
 import braille.display.gesture
 import brailleInput
@@ -151,7 +152,7 @@ class BrailleDisplayDriver(braille.display.driver.BrailleDisplayDriver):
 
 	@classmethod
 	def getManualPorts(cls):
-		return braille.getSerialPorts()
+		return braille.display.getSerialPorts()
 
 	def __init__(self, port="auto"):
 		super(BrailleDisplayDriver, self).__init__()
