+++ source/brailleDisplayDrivers/brailliantB.py
@@ -8,6 +8,7 @@ from typing import List, Union
 
 import serial
 import braille
+import braille.display
 import braille.display.driver
 import braille.display.gesture
 import inputCore
@@ -157,7 +158,7 @@ class BrailleDisplayDriver(braille.display.driver.BrailleDisplayDriver):
 
 	@classmethod
 	def getManualPorts(cls):
-		return braille.getSerialPorts()
+		return braille.display.getSerialPorts()
 
 	def __init__(self, port="auto"):
 		super(BrailleDisplayDriver, self).__init__()
