+++ source/brailleDisplayDrivers/nattiqbraille.py
@@ -7,6 +7,7 @@
 import serial
 import bdDetect
 import braille
+import braille.display
 import braille.display.driver
 import braille.display.gesture
 import inputCore
@@ -49,7 +50,7 @@ class BrailleDisplayDriver(braille.display.driver.BrailleDisplayDriver):
 
 	@classmethod
 	def getManualPorts(cls):
-		return braille.getSerialPorts()
+		return braille.display.getSerialPorts()
 
 	def __init__(self, port="auto"):
 		super(BrailleDisplayDriver, self).__init__()
