+++ source/brailleDisplayDrivers/superBrl.py
@@ -8,6 +8,7 @@ from typing import List
 import serial
 import bdDetect
 import braille
+import braille.display
 import braille.display.driver
 import hwIo
 from hwIo import intToByte
@@ -43,7 +44,7 @@ class BrailleDisplayDriver(braille.display.driver.BrailleDisplayDriver):
 
 	@classmethod
 	def getManualPorts(cls):
-		return braille.getSerialPorts()
+		return braille.display.getSerialPorts()
 
 	def __init__(self, port="Auto"):
 		super(BrailleDisplayDriver, self).__init__()
