+++ tests/unit/test_braille/test_windowBrailleCells.py
@@ -9,11 +9,12 @@ import unittest
 
 import braille
 import braille.buffers
+import braille.display
 
 
-def _getDisplayDimensions(dimensions: braille.DisplayDimensions) -> braille.DisplayDimensions:
+def _getDisplayDimensions(dimensions: braille.display.DisplayDimensions) -> braille.display.DisplayDimensions:
 	"""Used to build a braille handler with particular dimensions."""
-	return braille.DisplayDimensions(
+	return braille.display.DisplayDimensions(
 		numRows=2,
 		numCols=20,
 	)
